+++ Source/WebKit/UIProcess/Extensions/glib/WebExtensionGLib.cpp
@@ -63,17 +63,6 @@ WebExtension::WebExtension(GFile* resourcesFile, RefPtr<API::Error>& outError)
     }
 }
 
-WebExtension::WebExtension(const JSON::Value& manifest, Resources&& resources)
-    : m_manifestJSON(manifest)
-    , m_dataResources(toDataResources(resources))
-    , m_stringResources(toStringResources(resources))
-{
-    auto manifestString = manifest.toJSONString();
-    RELEASE_ASSERT(manifestString);
-
-    m_stringResources.set("manifest.json"_s, manifestString);
-}
-
 Expected<Ref<API::Data>, RefPtr<API::Error>> WebExtension::resourceDataForPath(const String& originalPath, CacheResult cacheResult, SuppressNotFoundErrors suppressErrors)
 {
     ASSERT(originalPath);
