+++ Source/WebKit/WebProcess/Extensions/Bindings/JSWebExtensionWrapper.h
@@ -209,6 +209,11 @@ JSValueRef toWindowObject(JSContextRef, WebPage&);
 
 RefPtr<JSON::Value> fromJSValue(JSContextRef, JSValueRef);
 RefPtr<JSON::Value> toJSONValue(JSContextRef, JSValueRef, NullValuePolicy = NullValuePolicy::NotAllowed, ValuePolicy = ValuePolicy::Recursive);
+inline JSValueRef toJSValueRefOrJSNull(JSContextRef context, RefPtr<JSON::Object> value)
+{
+    ASSERT(context);
+    return value ? fromJSON(context, value) : JSValueMakeNull(context);
+}
 
 #ifdef __OBJC__
