+++ Source/WebKit/WebProcess/Extensions/Bindings/Scripts/IDLAttributes.json
@@ -23,6 +23,10 @@
             "values": ["*"],
             "supportsConjunction": true
         },
+        "Platform": {
+            "contextsAllowed": ["attribute"],
+            "values": ["*"]
+        },
         "CallbackHandler": {
             "contextsAllowed": ["argument"]
         },
